XD-2358-2359-2360-2549
- Upgrade to Spring Integration Kafka 1.0.0.M3
- Replace the Kafka Source with one based on the KafkaMessageDrivenChannelAdapter
- Add support for listening to a subset of the original set of partitions ('partitions' module parameter)
- Evenly split the listened partitions between module instance and coordinate via Zookeeper
- Add support for starting reading from a series of initial offsets ('replay')
- Pluggable strategy for storing offsets during consumption (supported: in-memory/transient and Redis)
