--- extensions/spring-xd-extension-kafka/src/main/java/org/springframework/integration/x/kafka/KafkaSourceOffsetOptionsMetadata.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-
-package org.springframework.integration.x.kafka;
-
-import org.springframework.xd.module.options.spi.ModuleOption;
-
-/**
- * @author Marius Bogoevici
- */
-public class KafkaSourceOffsetOptionsMetadata {
-
-	private String offsetStorage = "inmemory";
-
-	public String getOffsetStorage() {
-		return offsetStorage;
-	}
-
-	@ModuleOption("offset storage strategy")
-	public void setOffsetStorage(String offsetStorage) {
-		this.offsetStorage = offsetStorage;
-	}
-
-}
